Conflicts:
	robot/simulator/controllers/cozmo_c_controller/CMakeLists.txt
	robot/simulator/controllers/cozmo_docking_controller_test/CMakeLists.txt

+++ coretech/common/include/anki/common/robot/matlabConverters.h
@@ -126,7 +126,7 @@ namespace Anki {
         return Array<Type>();
       }
 
-      Array<Type> array = AllocateArrayFromHeap<Type>(dimensions[0], dimensions[1]);
+      Array<Type> array = AllocateArrayFromHeap<Type>(static_cast<s32>(dimensions[0]), static_cast<s32>(dimensions[1]));
 
       for(mwSize y=0; y<dimensions[0]; ++y) {
         Type * const pArray = array.Pointer(static_cast<s32>(y), 0);
